Post review fixes
